refactor(middleware): extract shared macro utils, justify casts
