+++ zio-http-core/shared/src/main/scala-3/zio/http/ContextHandlerMacro.scala
@@ -41,13 +41,8 @@ private[http] object ContextHandlerMacro {
     }
   }
 
-  private def findIsNominal(using q: Quotes)(t: q.reflect.TypeRepr): q.reflect.Term = {
-    import q.reflect.*
-    Implicits.search(TypeRepr.of[IsNominalType].appliedTo(List(t))) match {
-      case s: ImplicitSearchSuccess => s.tree
-      case _                        => report.errorAndAbort(s"Cannot find IsNominalType for ${t.show}.")
-    }
-  }
+  private def findIsNominal(using q: Quotes)(t: q.reflect.TypeRepr): q.reflect.Term =
+    MacroUtils.findIsNominal(using q)(t)
 
   private def genHandler[H: Type](h: Expr[H], ctxTypesRaw: List[Any])(using q: Quotes): Expr[Handler[?, ?]] = {
     import q.reflect.*
@@ -84,25 +79,14 @@ private[http] object ContextHandlerMacro {
       case Nil       =>
         val args: List[Term] = (req.asTerm :: acc.map(_.asTerm)).asInstanceOf[List[Term]]
         val fnTpe            = fnTerm.tpe.widen
-        val callTerm: Term   =
-          if (fnTpe.typeSymbol.fullName == "scala.FunctionXXL") {
-            val iarrayTpe  = TypeRepr.of[IArray[Any]]
-            val consSym    = iarrayTpe.typeSymbol.companionModule.methodMember("apply").head
-            val consSelect = Select(Ref(iarrayTpe.typeSymbol.companionModule), consSym)
-            val iarrayCall = Apply(consSelect, args)
-            val applyXXL   = fnTpe.typeSymbol.methodMember("apply").head
-            Apply(Select(fnTerm, applyXXL), List(iarrayCall))
-          } else {
-            val applySym = fnTpe.typeSymbol.methodMember("apply").head
-            Apply(Select(fnTerm, applySym), args)
-          }
+        val callTerm: Term = MacroUtils.buildFunctionCall(using q)(fnTerm, args)
         callTerm.asExprOf[Response | Halt]
       case t :: rest =>
         t.asType match {
           case '[tpe] =>
             val evTerm                           = findIsNominal(using q)(t)
             val evExpr: Expr[IsNominalType[tpe]] = evTerm.asExprOf[IsNominalType[tpe]]
-            val g: Expr[tpe]                     = '{ $ctx.asInstanceOf[Context[tpe]].get[tpe](using $evExpr) }
+            val g: Expr[tpe]                     = '{ $ctx.asInstanceOf[Context[tpe]].get[tpe](using $evExpr) } // cast sound: IsNominalType evidence
             loop(rest, acc :+ g.asExprOf[Any])
         }
     }
